backend/requirements.txt
Flask==3.1.3
Flask-Cors==6.0.1
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.48
psycopg[binary]==3.2.12
PyJWT==2.10.1
Werkzeug==3.1.3
python-dotenv==1.1.1
